fix: address review round 1 (docs link, pure-C consumer, uint32 casts)
- docs/C_ABI.md: correct the README relative-link depth (../../ -> ..); the file
  is one level deep under docs/, so ../../README.md resolved outside the repo.
- tests/consumer: make C++ optional (project LANGUAGES C + check_language(CXX),
  consumer_hpp gated behind CXX availability), so the smoke project also
  configures with no C++ compiler present -- actually exercising the "pure-C
  consumers work" acceptance criterion rather than only the C++-present path.
- tests/consumer/main.{c,cpp}: cast sw_api_version() (uint32_t) to unsigned for
  %u, matching examples/c/sw_dump.c and avoiding format-mismatch UB/warnings.

Addresses Copilot inline comments and the round-1 review on #18.

+++ tests/consumer/main.c
@@ -6,6 +6,7 @@
 #include <stdio.h>
 
 int main(void) {
-    printf("sensorwatch C consumer OK, ABI=%u\n", sw_api_version());
+    /* Cast for %u: sw_api_version() returns uint32_t (matches examples/c/sw_dump.c). */
+    printf("sensorwatch C consumer OK, ABI=%u\n", (unsigned)sw_api_version());
     return 0;
 }
